equity_pipeline/requirements.txt
supabase==2.7.4
pandas==2.2.2
yfinance==0.2.28
requests==2.32.3
python-dotenv==1.0.1
edgartools==2.23.0
google-genai==0.8.0